Change TimeSpan to use decimals to not lose precision during deserialization

+++ src/ServiceStack.Text/Support/TimeSpanConverter.cs
@@ -54,7 +54,7 @@ namespace ServiceStack.Text.Support
             int days = 0;
             int hours = 0;
             int minutes = 0;
-            double seconds = 0;
+            decimal seconds = 0;
             int sign = 1;
 
             if (xsdDuration.StartsWith("-", StringComparison.Ordinal))
@@ -95,13 +95,13 @@ namespace ServiceStack.Text.Support
                 string[] s = m[m.Length - 1].SplitOnFirst('S');
                 if (s.Length == 2)
                 {
-                    double millis;
-                    if (double.TryParse(s[0], out millis))
+                    decimal millis;
+                    if (decimal.TryParse(s[0], out millis))
                         seconds = millis;
                 }
             }
 
-            double totalSecs = 0
+            decimal totalSecs = 0
                     + (days * 24 * 60 * 60)
                     + (hours * 60 * 60)
                     + (minutes * 60)
